Replace ButtonStyle by AppCompatStyles
It turns out it is not slower to apply all the styles to a theme with
a "fatter" theme.

Also, programmatic styles are slower, because code for non existing
xml style is still run, so programmatic style applying only adds up to
execution time. Programmatic styles are therefore only interesting for
views that don't run the xml style logic when not needed, which is
unfortunately not the case for AppCompat Views.