[Failed] Preferences integration with Obj-C KVO
This has been committed because it demonstrates how to add Objective-C
source code to a Kotlin/Native library, and might be used in the future
to wrap Obj-C Key/Value Observing for other purposes (e.g. integration
with kotlinx.coroutines Flow).

Integration into Preferences failed because of several reasons:
- Having weak references for the wrapper to follow the Android API was
impractical.
- Obj-C KVO doesn't allow to observe all the keys of an object, and
knowing all the keys would pollute the implementation of the Kotlin
delegates in the common code.
- There could have been threading issues as the observer could have been
notified on any thread, which would have required to redispatch again,
making the implementation even more complex.