Add Auto-Manifest plugin
Auto-Manifest automatically generates AndroidManifest packageName based on Gradle module structure. Since Splitties has `modules` and `fun-packs` separated, it would put `modules` in the middle of the package. That's why the plugin is applied under `modules/build.gradle.kts`. This made me realize a small issue on the plugin and released a fix.

Once change required in the code was that `views-dsl` used `views.dsl.core` as packageName and I had to change the imports. Since these imports are not exposed, I don't think it is a breaking change.