Introduce AndroidX DataStore backed storage in Preferences
This commit also renames the SharedPreferences interface and
other symbols into PreferencesStorage to avoid any future confusion.

This breaks the API a bit, but it's binary compatible on Android,
and it's unlikely this API surface was actually in use anyway,
since it was there mostly to allow code sharing,
internally to Splitties, and not part of the added value of the
Preferences split.