Protect race from potential rare race condition
This commit simplifies the implementation of race,
all while making it immune to race conditions.

Such a thing could potentially happen if a racer was
launched after one won, on a separate thread,
before the `raceWon` boolean change was visible to that thread.
Consequence would be needlessly delaying the race end after
that last racer, despite the race having already been won.

Extremely unlikely, and probably never happened in the wild,
but now, it's impossible.